before the defense

+++ final_project/src/main/java/by/it_academy/jd2/food_control/service/ProductService.java
@@ -8,6 +8,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
+import javax.persistence.OptimisticLockException;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -37,8 +38,8 @@ public class ProductService implements IService<Product, Long> {
 
     @Override
     public List<Product> findAll(SearchFilter filter) {
-        int offset = (int) filter.getPage();
-        int limit = (int) filter.getSize();
+        int offset = Math.toIntExact(filter.getPage());
+        int limit = Math.toIntExact(filter.getSize());
 
         try {
             return this.repository.findByName(filter.getName(), PageRequest.of(offset, limit));
@@ -51,7 +52,9 @@ public class ProductService implements IService<Product, Long> {
     public boolean deleteId(Long id, Long version) {
         try {
             Product product = findById(id);
-            product.setVersion(version);
+            if (product.getVersion() > version) {
+                throw new OptimisticLockException();
+            }
             this.repository.delete(product);
             return true;
         } catch (Exception e) {
@@ -60,24 +63,36 @@ public class ProductService implements IService<Product, Long> {
     }
 
     @Override
-    public Product updateItem(Long id, Product item) {
+    public Product updateItem(Long id, Product item, Long version) {
         Product product = null;
         try {
             product = this.repository.findById(id).get();
             product.setName(item.getName());
             product.setBrand(item.getBrand());
             product.setCalories(item.getCalories());
+
+//            testVersion(id);
+
             product.setProtein(item.getCalories());
             product.setProtein(item.getProtein());
             product.setFats(item.getFats());
             product.setCarbohydrates(item.getCarbohydrates());
             product.setMeasure(item.getMeasure());
             product.setUpdateDate(LocalDateTime.now());
-            product.setVersion(item.getVersion());
+
+            if (product.getVersion() > version) {
+                throw new OptimisticLockException();
+            }
             return this.repository.save(product);
         } catch (Exception e) {
             throw new IllegalArgumentException("Не удалось обновить продукт");
         }
+    }
+
+    private void testVersion(Long id) {
+        Product product1 = this.repository.findById(id).get();
+        product1.setName("update");
+        this.repository.save(product1);
     }
 
     @PostConstruct
